Renamed folders and classes to adhere to Java naming conventions
