+++ src/main/java/com/softserve/teachua/controller/test/QuestionsController.java
@@ -6,8 +6,8 @@ import com.softserve.teachua.dto.test.question.ImportProfile;
 import com.softserve.teachua.dto.test.question.QuestionDatabaseResponse;
 import com.softserve.teachua.dto.test.question.QuestionPreview;
 import com.softserve.teachua.dto.test.question.QuestionResponse;
-import com.softserve.teachua.dto.test.question.questionExcel.ExcelQuestionParsingResponse;
-import com.softserve.teachua.dto.test.question.questionExcel.QuestionDataRequest;
+import com.softserve.teachua.dto.test.question.question_excel.ExcelQuestionParsingResponse;
+import com.softserve.teachua.dto.test.question.question_excel.QuestionDataRequest;
 import com.softserve.teachua.model.test.QuestionCategory;
 import com.softserve.teachua.model.test.QuestionType;
 import com.softserve.teachua.security.UserPrincipal;
