+++ src/main/java/com/softserve/teachua/dto/certificate_excel/ExcelParsingResponse.java
@@ -1,6 +1,6 @@
 package com.softserve.teachua.dto.certificate_excel;
 
-import com.softserve.teachua.dto.databaseTransfer.ExcelParsingMistake;
+import com.softserve.teachua.dto.database_transfer.ExcelParsingMistake;
 import java.util.ArrayList;
 import java.util.List;
 import lombok.Data;
