--- src/main/java/com/softserve/teachua/dto/databaseTransfer/ExcelLoadSuccess.java
@@ -1,4 +1,0 @@
-package com.softserve.teachua.dto.databaseTransfer;
-
-public class ExcelLoadSuccess {
-}
