+++ src/main/java/com/softserve/teachua/dto/database_transfer/ExcelConvertToFormatStringContactsData.java
@@ -1,4 +1,4 @@
-package com.softserve.teachua.dto.databaseTransfer;
+package com.softserve.teachua.dto.database_transfer;
 
 import org.springframework.stereotype.Component;
