+++ src/main/java/com/softserve/teachua/dto/database_transfer/ExcelLoadSuccess.java
@@ -1,0 +1,4 @@
+package com.softserve.teachua.dto.database_transfer;
+
+public class ExcelLoadSuccess {
+}
