+++ src/main/java/com/softserve/teachua/dto/database_transfer/model/ClubExcel.java
@@ -1,4 +1,4 @@
-package com.softserve.teachua.dto.databaseTransfer.model;
+package com.softserve.teachua.dto.database_transfer.model;
 
 import com.softserve.teachua.dto.marker.Convertible;
 import java.util.List;
