+++ src/main/java/com/softserve/teachua/service/impl/ContactTypeServiceImpl.java
@@ -3,9 +3,9 @@ package com.softserve.teachua.service.impl;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.softserve.teachua.converter.DtoConverter;
-import com.softserve.teachua.dto.contactType.ContactTypeProfile;
-import com.softserve.teachua.dto.contactType.ContactTypeResponse;
-import com.softserve.teachua.dto.contactType.SuccessCreatedContactType;
+import com.softserve.teachua.dto.contact_type.ContactTypeProfile;
+import com.softserve.teachua.dto.contact_type.ContactTypeResponse;
+import com.softserve.teachua.dto.contact_type.SuccessCreatedContactType;
 import com.softserve.teachua.exception.AlreadyExistException;
 import com.softserve.teachua.exception.DatabaseRepositoryException;
 import com.softserve.teachua.exception.NotExistException;
