+++ src/main/java/com/softserve/teachua/service/impl/DataLoaderServiceImpl.java
@@ -4,14 +4,14 @@ import com.softserve.teachua.dto.category.CategoryProfile;
 import com.softserve.teachua.dto.center.CenterProfile;
 import com.softserve.teachua.dto.center.SuccessCreatedCenter;
 import com.softserve.teachua.dto.club.ClubProfile;
-import com.softserve.teachua.dto.databaseTransfer.ExcelConvertToFormatStringContactsData;
-import com.softserve.teachua.dto.databaseTransfer.ExcelParsingData;
-import com.softserve.teachua.dto.databaseTransfer.model.CategoryExcel;
-import com.softserve.teachua.dto.databaseTransfer.model.CenterExcel;
-import com.softserve.teachua.dto.databaseTransfer.model.ClubExcel;
-import com.softserve.teachua.dto.databaseTransfer.model.DistrictExcel;
-import com.softserve.teachua.dto.databaseTransfer.model.LocationExcel;
-import com.softserve.teachua.dto.databaseTransfer.model.StationExcel;
+import com.softserve.teachua.dto.database_transfer.ExcelConvertToFormatStringContactsData;
+import com.softserve.teachua.dto.database_transfer.ExcelParsingData;
+import com.softserve.teachua.dto.database_transfer.model.CategoryExcel;
+import com.softserve.teachua.dto.database_transfer.model.CenterExcel;
+import com.softserve.teachua.dto.database_transfer.model.ClubExcel;
+import com.softserve.teachua.dto.database_transfer.model.DistrictExcel;
+import com.softserve.teachua.dto.database_transfer.model.LocationExcel;
+import com.softserve.teachua.dto.database_transfer.model.StationExcel;
 import com.softserve.teachua.dto.district.DistrictProfile;
 import com.softserve.teachua.dto.location.LocationProfile;
 import com.softserve.teachua.dto.station.StationProfile;
