+++ src/test/java/com/softserve/teachua/TestUtils.java
@@ -45,8 +45,8 @@ import com.softserve.teachua.dto.certificate.CertificateVerificationResponse;
 import com.softserve.teachua.dto.certificate_by_template.CertificateByTemplateTransfer;
 import com.softserve.teachua.dto.certificate_excel.CertificateExcel;
 import com.softserve.teachua.dto.googleapis.QuizResult;
-import com.softserve.teachua.dto.test.answer.answerExcel.AnswerExcel;
-import com.softserve.teachua.dto.test.question.questionExcel.QuestionExcel;
+import com.softserve.teachua.dto.test.answer.answer_excel.AnswerExcel;
+import com.softserve.teachua.dto.test.question.question_excel.QuestionExcel;
 import com.softserve.teachua.dto.user.UserLogin;
 import com.softserve.teachua.model.Certificate;
 import com.softserve.teachua.model.CertificateDates;
