+++ src/test/java/com/softserve/teachua/service/test/QuestionExcelServiceTest.java
@@ -6,10 +6,10 @@ import static com.softserve.teachua.TestUtils.getEmptySheet;
 import static com.softserve.teachua.TestUtils.getIndexes;
 import static com.softserve.teachua.TestUtils.getListOfEmptyCells;
 import static com.softserve.teachua.TestUtils.getQuestionExcel;
-import com.softserve.teachua.dto.databaseTransfer.ExcelParsingMistake;
-import com.softserve.teachua.dto.test.answer.answerExcel.AnswerExcel;
-import com.softserve.teachua.dto.test.question.questionExcel.ExcelQuestionParsingResponse;
-import com.softserve.teachua.dto.test.question.questionExcel.QuestionExcel;
+import com.softserve.teachua.dto.database_transfer.ExcelParsingMistake;
+import com.softserve.teachua.dto.test.answer.answer_excel.AnswerExcel;
+import com.softserve.teachua.dto.test.question.question_excel.ExcelQuestionParsingResponse;
+import com.softserve.teachua.dto.test.question.question_excel.QuestionExcel;
 import com.softserve.teachua.service.ExcelParserService;
 import com.softserve.teachua.service.test.impl.QuestionExcelServiceImpl;
 import com.softserve.teachua.constants.excel.ExcelColumn;
